@@ -1,24 +1,28 @@
 name: appengine
-version: 0.12.1
+version: 0.13.0
 description: >-
   Support for using Dart as a custom runtime on
   Google App Engine Flexible Environment
 homepage: https://github.com/dart-lang/appengine
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  fixnum: ^0.10.0
+  fixnum: ^1.0.0
   gcloud: ^0.7.0
-  googleapis_auth: ^0.2.12
-  grpc: ^2.2.0
-  http: ^0.12.0
-  logging: ^0.11.0
-  path: ^1.6.0
-  protobuf: ^1.0.0
-  stack_trace: ^1.9.0
+  googleapis_auth: ^1.1.0
+  grpc: ^3.0.0
+  http: ^0.13.3
+  logging: ^1.0.1
+  path: ^1.8.0
+  protobuf: ^2.0.0
+  stack_trace: ^1.10.0
 
 dev_dependencies:
-  test: ^1.6.0
-  pedantic: ^1.8.0
+  test: ^1.17.5
+  pedantic: ^1.11.0
+
+dependency_overrides:
+  gcloud:
+    path: ../gcloud
